refactor(graph): make ReplaceNode self-contained; audit checkpoint_every type; update docs
GraphRewrite changes:
- Reorder method declarations in header/impl: RemoveNode before
  ReplaceNode (callers read top-down: remove is the simpler operation).
- Remove the 'if empty -> RemoveNode' shortcut from ReplaceNode.
  ReplaceNode now unconditionally validates and stores replacements.
  Empty-list ReplaceNode is valid but asserts that the caller intended
  ReplaceNode (not RemoveNode), making the two operations orthogonal.
  Commit() already handles empty replacements as a no-op.

GraphPassManager:
- Change checkpoint_every from size_t to uint32_t in both PassContext
  and SetCheckpointEvery(). size_t is architecture-dependent; uint32_t
  is more than sufficient for pass counts and avoids implicit narrowing
  in cross-platform contexts.

Documentation (model_graph_design_v2.md):
- Note that Commit() is const (reflects current implementation).
- Add validation/materialization separation status note.
- Add all-or-nothing error semantics and empty pipeline behavior.
- Add fusion flags usage convention section.
- Add [已实现]/[未实现]/[规划中] status tags to pass roadmap.
- Update PassContext delayed item and add rule 10 implementation note.

+++ include/aethermind/model/graph/graph_pass_manager.h
@@ -13,7 +13,7 @@ namespace aethermind {
 
 struct PassContext {
     uint32_t opt_level = 2;
-    size_t checkpoint_every = 0;
+    uint32_t checkpoint_every = 0;
 
     bool enable_qkv_fusion = true;
     bool enable_swiglu_fusion = true;
@@ -35,7 +35,7 @@ public:
     explicit GraphPassManager(PassContext ctx) noexcept;
 
     GraphPassManager& Add(std::unique_ptr<GraphPass> pass);
-    GraphPassManager& SetCheckpointEvery(size_t pass_count) noexcept;
+    GraphPassManager& SetCheckpointEvery(uint32_t pass_count) noexcept;
 
     AM_NODISCARD StatusOr<ModelGraph> Run(const ModelGraph& graph) const;
